LangChain Chatbot/requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-openai>=0.0.2
python-dotenv>=1.0.0
pypdf>=3.17.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
beautifulsoup4>=4.12.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tiktoken>=0.5.1
openai>=1.12.0
pydantic>=2.0.0
typing-extensions>=4.7.0
